refactor(components/taigaimagesmodal): move modal code to a component
