+++ src/components/TaigaImagesModal/styles.ts
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+export const ModalBody = styled.div`
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+  justify-items: center;
+  gap: 16px;
+`
