+++ src/pages/Home/styles.ts
@@ -11,25 +11,6 @@ export const Container = styled.div`
   align-items: center;
 `
 
-export const ModalBody = styled.div`
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: center;
-  justify-items: center;
-  gap: 16px;
-`
-
-export const Image = styled.div`
-  width: 240px;
-  height: 240px;
-  border-radius: 50%;
-  background-position: 50% 50%;
-  background-size: cover;
-  background-repeat: no-repeat;
-  cursor: pointer;
-`
-
 export const LeftContent = styled.div`
   width: 360px;
   height: 100%;
