+++ src/styles/GlobalComponents.ts
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+export const Image = styled.div`
+  width: 240px;
+  height: 240px;
+  border-radius: 50%;
+  background-position: 50% 50%;
+  background-size: cover;
+  background-repeat: no-repeat;
+  cursor: pointer;
+`
